+++ yash-syntax/src/parser/error.rs
@@ -159,6 +159,8 @@ pub enum SyntaxError {
     InvalidFunctionBody,
     /// The keyword `in` is used as a command name.
     InAsCommandName,
+    /// The keyword `]]` is used as a command name.
+    CloseBracketBracketAsCommandName,
     /// A pipeline is missing after a `&&` or `||` token.
     MissingPipeline(AndOr),
     /// Two successive `!` tokens.
@@ -288,8 +290,14 @@ impl SyntaxError {
                 "the delimiter to close the here-document content is missing"
             }
             UnclosedArrayValue { .. } => "the array assignment value is not closed",
-            UnopenedGrouping | UnopenedSubshell | UnopenedLoop | UnopenedDoClause | UnopenedIf
-            | UnopenedCase | InAsCommandName => "the compound command delimiter is unmatched",
+            UnopenedGrouping
+            | UnopenedSubshell
+            | UnopenedLoop
+            | UnopenedDoClause
+            | UnopenedIf
+            | UnopenedCase
+            | InAsCommandName
+            | CloseBracketBracketAsCommandName => "the compound command delimiter is unmatched",
             UnclosedGrouping { .. } => "the grouping is not closed",
             EmptyGrouping => "the grouping is missing its content",
             UnclosedSubshell { .. } => "the subshell is not closed",
@@ -421,7 +429,9 @@ impl SyntaxError {
             UnopenedCase => "not in a `case` command",
             UnclosedCase { .. } => "expected `esac`",
             MissingFunctionBody | InvalidFunctionBody => "expected a compound command",
-            InAsCommandName => "cannot be used as a command name",
+            InAsCommandName | CloseBracketBracketAsCommandName => {
+                "cannot be used as a command name"
+            }
             DoubleNegation => "only one `!` allowed",
             BangAfterBar => "`!` not allowed here",
             RedundantToken => "unexpected token",